requirements.txt
Flask
scapy
pyyaml
pyteal 
py-algorand-sdk